Add operate() which calls a function to eval input
  The idea is that operate() can be called to evaluate whatever has
been input to the calculator so far. I have some ideas as for how it
could behave recursively, sort of like such:

if LHS and RHS are num, call one of the functions on them

if either LHS or RHS are string, call operate on them individually,
  reducing them.

  This requires some sort of string parsing option at the start of
operate which I will get to soon.